Use a unified manage module to manage the agents
